[ADD] multi db connections

+++ src/elf/rand.cpp
@@ -142,7 +142,7 @@ void roll_rm(const roll_req &req, roll_res &res, int times, bool weight)
     roll_res::iterator itr_d;
     int weight_sum = 0;
 
-    assert(req.size() >= times);
+    assert(req.size() >= (size_t)times);
 
     if (weight) {
         for (; itr_c != req.end(); ++itr_c) {
